New command line option --update-old-devices to update the serial number of old servers
Add an alternative method to resolve server serial number and name
conflicts which updates the serial number of the existing server object
in Netbox with the same name as the current server.

This option addresses the situation as --purge-old-devices but does
not delete the existing server object in Netbox.

This option is useful if a server is replaced by another with the same
name and it is desired to keep all its properties that have been added
manually in Netbox, so the existing server object will be updated with
the serial number of the new server.

+++ netbox_agent/server.py
@@ -253,12 +253,16 @@ class ServerBase():
         )
         return new_blade
 
-    def _netbox_deduplicate_server(self):
+    def _netbox_deduplicate_server(self, purge):
         serial = self.get_service_tag()
         hostname = self.get_hostname()
         server = nb.dcim.devices.get(name=hostname)
         if server and server.serial != serial:
-            server.delete()
+            if purge:
+                server.delete()
+            else:
+                server.serial = serial
+                server.save()
 
     def _netbox_create_server(self, datacenter, tenant, rack):
         device_role = get_device_role(config.device.server_role)
@@ -384,8 +388,11 @@ class ServerBase():
         rack = self.get_netbox_rack()
         tenant = self.get_netbox_tenant()
 
+        if config.update_old_devices:
+            self._netbox_deduplicate_server(purge=False)
+
         if config.purge_old_devices:
-            self._netbox_deduplicate_server()
+            self._netbox_deduplicate_server(purge=True)
 
         if self.is_blade():
             chassis = nb.dcim.devices.get(
